+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/actions/ShowSettingsAction.kt
@@ -16,12 +16,15 @@
 
 package it.czerwinski.intellij.wavefront.editor.actions
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.options.ShowSettingsUtil
 import it.czerwinski.intellij.wavefront.settings.WavefrontObjSettingsConfigurable
 
 class ShowSettingsAction : AnAction() {
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
 
     override fun update(event: AnActionEvent) {
         event.presentation.isEnabled = true