+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/actions/ToggleCropTexturesAction.kt
@@ -16,12 +16,15 @@
 
 package it.czerwinski.intellij.wavefront.editor.actions
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.Toggleable
 import com.intellij.openapi.project.DumbAware
 import it.czerwinski.intellij.wavefront.editor.model.ShadingMethod
 
 class ToggleCropTexturesAction : GLPreviewFileEditorAction(), DumbAware, Toggleable {
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
 
     override fun update(event: AnActionEvent) {
         val editor = findGLPreviewFileEditor(event)